Fix marked ayah of parah page
It was not updating selected state of mutashabiha ayats correctly. To
fix this,
- Remove selection state from main ParaAyatModel and Ayat
- Selection state now lives in the MarkedAyahsPage where its needed
- Also fix that ReadQuranPage doesn't update after ayahs were removed
from the model

The new code is simper but also repaints the whole page when we tap
or untap an ayah in MarkedAyahsPage to select/unselect it. But this is
not something that is expected to be done a lot so for now I think this
will be enough.