Fixes for reflow mode
Unfortunately flutter doesn't allow me to have a different height
for each page. Thus I need to pre specify a height that will work
for all pages. To work around this,

1. Have a specific set of font sizes
2. Have a multiplier for each font size

We then multiply the calculated height by the multiplier and _hope_
that it works for all pages.

Since the pages will be scrollable on bigger font sizes, reset the
vertical scroll of top level scrollable to 0 once the page changes