Properly fix the availableHeight
I was using View.of(context).viewPadding.top to get the top padding.
This returned a raw non-logical pixel value on android and thus
the content fit on a lot of screens without any scroll. But not all
screens. There were weird bugs. On desktop, it always caused a scroll
unless I manually subtracted appBar height.

Subtracting appbar height makes sense, as the content is occupying
space below the appbar. To properly fix this, divide the viewPadding
by dpr so that we get logical pixel value.

Verfied by checking MediaQuery.paddingOf(context) before Scaffold and
the value is same.