Introduce new font for 16 line quran
The new font is Digital Khatt and the text is taken from Quranic
Universal Library and then adapted for 16 line page format.

This text includes ayah end markers so the logic is changed a little
bit to work with that. There is a slight issue with tapping ruku marker
with this change but I haven't done much about it yet.