Rework how text data is stored
Instead of storing the data in assets, store it directly in the code.
This allows us to load it much faster.

Also, store just a list of ayahs. And store layout information
separately. This allows us to use different layouts without duplicating
the text data

In prepration for 15 line indopak