Flatten ParaModel
- Store ayats directly instead of by para
- Page view now inits once with the whole quran. We load the page
that is going to be shown on demand.

The main benefit is reduced complexity as we now don't need to do
special tricks to make PageController work correctly when jumping
paras. This also improves performance, there is no lag when switching
para anymore because its not really a switch, we just jump to a
certain page.