Revert to previous way of data backup
New way is more complex and not sure worth it for now. The main issue
is that I dont have permission to delete file that I did not create.
This means that if the user deleted my app and then installed it again
, the app will not be able to delete/overwrite old backup files. This
can be worked around by keeping 1 backup file per install but that
feels clunky.